Fix Descriptor token assignment for non-exported methods/fields, rewrite ExportFileWriter for format 2.1
Descriptor tokens (JCVM §6.14 Table 6-39):
- All private methods (static + instance) now get token 0xFF
- Package-private static methods and constructors get token 0xFF
- Static initializers (<clinit>) get token 0xFF
- Non-exported static fields get token 0xFF

TokenAssigner:
- Exclude non-exported methods/fields from token map
- Skip compile-time constant fields (inlined by javac)

ExportFileWriter rewrite for Oracle-compatible format 2.1:
- CP starts at index 0 (not 1), tokens as u1 (not u2)
- CLASSREF entries with fully-qualified class names
- Super class chain in supers section
- Field attributes (u2 attribute_count + ConstantValue)
- Access flags: strip JVM ACC_SUPER (0x0020)
- Export format version from JavaCardVersion (2.1 for all versions)

API stubs:
- Util.arrayCompare() return type: short → byte
- APDU.getOffsetCdata() added (JC 3.0.5)

Oracle verifycap: 6/7 applets pass (CryptoApplet: pre-existing CP ordering issue)

+++ javacard-api/src/main/java/javacard/framework/Util.java
@@ -70,7 +70,7 @@ public class Util {
      * @param length  number of bytes to compare
      * @return 0 if equal, negative if src &lt; dest, positive if src &gt; dest
      */
-    public static short arrayCompare(byte[] src, short srcOff, byte[] dest, short destOff, short length) {
+    public static byte arrayCompare(byte[] src, short srcOff, byte[] dest, short destOff, short length) {
         return 0;
     }
